refactor, rename, and move getEmailHTML function to email notification provider
