+++ server/notifications/index.js
@@ -1,4 +1,4 @@
-import { EmailNotificationProvider } from './EmailNotificationProvider.js';
+import { EmailNotificationProvider } from './email/EmailNotificationProvider.js';
 
 export const providers = {
     email: EmailNotificationProvider
